% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/visualization.R
\name{axes}
\alias{axes}
\title{Add reference lines to a ggplot}
\usage{
axes(axis = "xy", xintercept = 0, yintercept = 0, color = "gray",
  size = 1)
}
\arguments{
\item{axis}{Character vector specifying the axis, either "x", "y", or "xy".}

\item{xintercept,}{yintercept Numeric value.}

\item{color}{Color.}

\item{size}{Size.}
}
\value{
A ggplot layer or layers.
}
\description{
This is a convenience function that creates x and/or y axis reference lines
that can be added to a ggplot object.
}

