% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biovars.R
\name{biovariables}
\alias{biovariables}
\title{Calculate the 19 bioclimatic variables.}
\usage{
biovariables(x)
}
\arguments{
\item{x}{Numeric vector of length 36 or 42. If calculating biovars for a
climate normal, should be 1 value per month for precip, tmax, and tmin,
arranged p1:p12 tx1:tx12 tn1:tn12. If calculating for a time series, uses
14 months per variable, with 13 and 14 being Jan-Feb of the next year. This
is automtically determined based on the length of the input vector.}
}
\value{
Numeric vector of length 19, representing the 19 bioclimatic variables
  in order.
}
\description{
This function calculates the 19 bioclimatic variables based on monthly precip
and monthly min and max temperatures. It is very similar to dismo::biovars
but differs in a couple key respects. First, it's much much faster. Second,
it only works on vectors, so to use it on rasters it should be paired with
raster::calc. Third, it works on time series in addition to normals, if fed
the right data. Fourth, Bio4 is not multiplied by 100 (dismo just does that
to save some resources). And finally, for Bios 8, 9, 18, and 19, if multiple
quarters have the same min or max value, this version computes the mean
rather than simply using the first one.
}

