% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spatial.R
\name{boundaryCrop}
\alias{boundaryCrop}
\title{Wrapper around loading and clipping by a shapefile.}
\usage{
boundaryCrop(raster, boundary.directory, boundary.filename)
}
\arguments{
\item{raster}{A *Raster object to be clipped.}

\item{boundary.directory}{The file folder path of the boundary shapefile.}

\item{boundary.filename}{The bare filename (no extension) of the shapefile.}
}
\value{
A clipped *Raster object.
}
\description{
This function loads a shapefile and then uses it to clip a raster object.
}

