% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spatial.R
\name{clp}
\alias{clp}
\alias{crp}
\title{Clip a raster.}
\usage{
clp(raster, polygon)
}
\arguments{
\item{raster}{A *Raster object to be clipped.}

\item{polygon}{A spatial polygon object used for clipping.}
}
\value{
A clipped *Raster object.
}
\description{
Convenience wrapper function that clips a raster object by a shapefile.
ASSUMES THE TWO OBJECTS HAVE MATCHING COORDINATE SYSTEMS.
}

