% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualization.R
\name{contourLevel}
\alias{contourLevel}
\title{Density level for a percentage contour}
\usage{
contourLevel(x, y, prob = 0.95)
}
\arguments{
\item{x,y}{Numeric vectors corresponding to scatterplot coordinates.}

\item{prob}{Vector of numerics between 0 and 1, representing the proportion of probability density to include in the contour.}
}
\value{
A vector of length equal to probs, specifying the breakpoints corresponding to each prob.
}
\description{
Get the 2D density breakpoint needed to generate a contour encompassing a
specific percentage of probability density.
}

