% Generated by roxygen2 (4.0.2): do not edit by hand
\name{crp}
\alias{clp}
\alias{crp}
\title{Clip a raster.}
\usage{
crp(raster, polygon)
}
\arguments{
\item{raster}{A *Raster object to be clipped.}

\item{polygon}{A spatial polygon object used for clipping.}
}
\value{
A clipped *Raster object.
}
\description{
Convenience wrapper function that clips a raster object by a shapefile.
ASSUMES THE TWO OBJECTS HAVE MATCHING COORDINATE SYSTEMS.
}

