% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cw}
\alias{cw}
\title{Build a colorwheel object}
\usage{
cw(data, xvar, yvar, resolution = 10, fade = "white")
}
\arguments{
\item{data}{A data frame.}

\item{xvar,}{yvar Names of data frame variables, as strings.}

\item{resolution}{Integer: one-dimensional granularity of raster legend.}

\item{fade}{Color at center of legend.}
}
\value{
A colorwheel object consisting of data, plots, and scales.
}
\description{
Generate colorwheel components
}

