% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colorwheel.R
\name{cwdata}
\alias{cwdata}
\title{Generate circular legend data.}
\usage{
cwdata(data, xvar, yvar, resolution = 10, origin = c(0, 0))
}
\description{
Construct a data frame to be used in color wheel legend plotting.
}

