% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getBoundary}
\alias{getBoundary}
\title{Load a shapefile.}
\usage{
getBoundary(directory, filename)
}
\arguments{
\item{directory}{The full path to the file folder where the shapefile lives, as a character.}

\item{filename}{The bare name of the shapefile, without extenstion, as a character.}
}
\value{
A spatial object.
}
\description{
This function loads a shapefile into memory for use in spatial or
 cartographic operations.
}

