% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{logit2prop}
\alias{logit2prop}
\title{Convert logit to proportion}
\usage{
logit2prop(x)
}
\arguments{
\item{x}{Numeric: logit.}
}
\value{
Numeric: proportion.
}
\description{
Logistic regression uses a logistic link function to convert between
proportions and the modeled dependent variable. This function back-converts,
for use when transforming model predictions into proportions.
}

