% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spatial.R
\name{motleyStack}
\alias{motleyStack}
\title{Stack rasters with differing extents.}
\usage{
motleyStack(x, intersect = F)
}
\arguments{
\item{x}{Layers to stack, as a vector of raster file paths or a list of
raster layers.}

\item{intersect}{Logical: Should only the minimum intersection of the input
layers be retained? (If FALSE, the default, layers will be extended to
ratain all data.)}
}
\value{
A raster stack.
}
\description{
The native raster::stack function fails if the input layers don't have
identical extents. This is a workaround that extends each input layer by a
buffer, and then intersects them. It will only work if all input layers share
other key spatial characteristics like resolution, projection, etc.
}

