% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{pairsData}
\alias{pairsData}
\title{Generate data for pairwise scatterplot matrix.}
\usage{
pairsData(data, xy_vars, z_vars = NULL, mirror = F)
}
\arguments{
\item{data}{Data frame, in "wide" format (i.e. must contain a column for each
xy_vars and z_vars).}

\item{xy_vars}{Character vector indicating variables to be paired (these
would be the x and y dimensions if the result is used for a scatterplot
matrix).}

\item{z_vars}{Optional character vector indicating additional variables to
retain for each record.}

\item{mirror}{Logical indicating whether to generate data for both upper and
lower trianges of the matrix (default is FALSE).}
}
\value{
A restructured data frame.
}
\description{
Convert a wide dataset into a format that can be fed to ggplot to create a
scatterplot matrix.
}

