% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spatial.R
\name{partitionRaster}
\alias{partitionRaster}
\title{Split a Raster* object into strips.}
\usage{
partitionRaster(raster, nchunks, chunk, subextent = NA)
}
\arguments{
\item{raster}{A *Raster object to be partitioned.}

\item{nchunks}{Integer representing number of partitions.}

\item{chunk}{Integer representing which of the paritions to create.}

\item{subextent}{Optional extent object indicating a subdomain of the raster
that should be partitioned.}
}
\value{
A cropped strip of the input raster object.
}
\description{
Partition a raster into vertical strips. Useful for parallel processing.
}

