% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colorwheel.R
\name{polarize}
\alias{polarize}
\title{Circular data transformation.}
\usage{
polarize(data, xvar, yvar, xyratio, xorigin = 0, yorigin = 0)
}
\description{
Add angle and distance variables to a data frame.
}

