% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spatial.R
\name{pwdSamplePA}
\alias{pwdSamplePA}
\title{Inverted pwdSample for presence-absence data.}
\usage{
pwdSamplePA(TP, TA, EP, EA, tr = 0.2, max_points = 1000)
}
\arguments{
\item{TP,}{TA, EP, EA Spatial points objects: Training and Evaluation,
Presence and Absence.}

\item{tr}{Numeric: maximum threshold difference between an eval point's
distance to nearest TA and TP to be considered similar.}

\item{max_points}{Integer: largest number of solutions to return for
evaluation presence and absence points.}
}
\value{
A list of two spatial points objects, one for presence evals and one
  for absence evals.
}
\description{
Select a subsample of evaluation presence and absence points, each of which
is equidistant to the nearest training presence and absence. This *may* help
remove bias generated by spatial autocorrelation in the same way that
dimso::pwdSample does for presence-only data.
}

