% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{sift}
\alias{sift}
\title{Subsetting by loop iterators.}
\usage{
sift(data, ...)
}
\arguments{
\item{data}{Data frame from which records will be selected.}

\item{...}{Filtration factors, saved as varialbes with names matching data
columns.}
}
\value{
Data frame subsetted to match filtration factors.
}
\description{
This convenience function is a wrapper around data frame subsetting, designed
for use inside loops that iterate through raster data directories.
}

