% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spatial.R
\name{stackBands}
\alias{stackBands}
\title{Stack up a single band from each of several multiband files.}
\usage{
stackBands(paths, band)
}
\arguments{
\item{paths}{A character vecor of raster filenames.}

\item{band}{A single integer or srting denoting which layer to grab.}
}
\value{
A raster stack with one layer from each file in paths.
}
\description{
This is a simple convenience function that pulls a named or numbered band
from each of several multiband files and stacks them up.
}

