% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spatial.R
\name{stackMatrix}
\alias{stackMatrix}
\title{Turn a data matrix into a raster stack.}
\usage{
stackMatrix(data, template)
}
\arguments{
\item{data}{A matrix.}

\item{template}{A raster layer to be used as spatial template; dims must
match data.}
}
\value{
A raster stack.
}
\description{
Create a raster stack from a matrix, where each matrix row is a pixel and
each column is a layer.
}

