% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spatial.R
\name{trimFast}
\alias{trimFast}
\title{Faster version of raster::trim.}
\usage{
trimFast(x, out = "raster")
}
\arguments{
\item{x}{Raster or matrix.}

\item{out}{Output file format, either "raster", or "matrix".}
}
\description{
Raster::trim can be insanely slow for large files. This alternative, modified
slightly from
http://blog.snap.uaf.edu/2012/09/10/speeding-up-r-raster-package-trim-function/,
runs way faster. But unlike the raster package version there's no guarantee
it won't run into memory limitations.
}

