% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convertRaster}
\alias{convertRaster}
\title{Convert raster files}
\usage{
convertRaster()
}
\description{
Convert raster file(s) to a new format. Via popup boxes, specify input
files, output file type, and output folder. All input files will be saved in
new format to the output folder, with the same filename as the original.
}

