% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/colorwheel.R
\name{cw}
\alias{cw}
\title{Build a colorwheel object}
\usage{
cw(data, xvar, yvar, resolution = 10, fade = "white", origin = c(0, 0),
  flip = NULL, left = left <- hsv(0.5, 1, 1), bottom = hsv(0.75, 1, 1),
  right = hsv(1, 1, 1), top = hsv(0.25, 1, 1))
}
\arguments{
\item{data}{A data frame.}

\item{xvar,}{yvar Names of data frame variables, as strings.}

\item{resolution}{Integer: one-dimensional granularity of raster legend.}

\item{fade}{Color at center of legend.}
}
\value{
A colorwheel object consisting of data, plots, and scales.
}
\description{
Generate colorwheel components
}

