% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{euclid}
\alias{euclid}
\title{Euclidian distance}
\usage{
euclid(z)
}
\arguments{
\item{z}{Numeric vector of z-scores or other distances.}
}
\description{
Calculate the Euclidian distance from a set of leg lengths. We use this to
calculate SED based on z-scores. Square root of sum of squares.
}

