% Generated by roxygen2 (4.0.2): do not edit by hand
\name{flake}
\alias{flake}
\title{Adjust coordinates for small multiples plotting.}
\usage{
flake(target, group, offset, summary = F)
}
\arguments{
\item{target}{Numeric vector of coordinates to be adjusted.}

\item{group}{Numeric vector of group factors, equal in length to target.}

\item{offset}{Numeric value representing offset distance per group.}

\item{summary}{Logical}
}
\description{
This function is used to progressively offset the coordinates of map layers
in long-format tabulated geographic data so that they can be plotted tightly
with overlapping extents. It is designed for use on multi-layer raster data
that has been converted into a data frame.
}
\details{
It returns a vector of adjusted coordinates equal in
length to the input coordinates.
}

