% Generated by roxygen2 (4.0.2): do not edit by hand
\name{frameRasters}
\alias{frameRasters}
\title{Create a data frame from raster files}
\usage{
frameRasters(metadata, layer_names = NA)
}
\arguments{
\item{metadata}{Data frame with a "path" variable and other metadata variables.}

\item{layer_names}{Character vector of layer names for multiband rasters (must be correct length).}
}
\value{
A data frame, with a row for each pixel of each raster in metadata.
}
\description{
This function takes a metadata frame (likely generated via parseMetadata) and
compiles all the files into a single data frame. Supports multiband files.
}

