% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{frameRasters}
\alias{frameRasters}
\title{Create a data frame from raster files}
\usage{
frameRasters(metadata, layer_names = NA, cpus = 1, bands = NA)
}
\arguments{
\item{metadata}{Data frame with a "path" variable and other metadata
variables.}

\item{layer_names}{Character vector of layer names for multiband rasters that
are missing layer names (must be correct length).}

\item{cpus}{Integer indicating number of CPUs to use for parallel processing.}

\item{bands}{Character vector denoting which layers to keep from multiband
files. Default is all.}
}
\value{
A data frame, with a row for each pixel of each raster in metadata.
}
\description{
This function takes a metadata frame (likely generated via parseMetadata) and
compiles all the files into a single data frame. Supports multiband files.
}

