% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spatial.R
\name{loadShapefile}
\alias{getBoundary}
\alias{loadShapefile}
\title{Load a shapefile.}
\usage{
loadShapefile(directory, filename)
}
\arguments{
\item{directory}{The full path to the file folder where the shapefile lives, as a character.}

\item{filename}{The bare name of the shapefile, without extenstion, as a character.}
}
\value{
A spatial object.
}
\description{
This function loads a shapefile into memory for use in spatial or
 cartographic operations.
}

